backend/requirements.txt
Django==3.2
django-filter==22.1
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.3
gunicorn==20.1.0